Changed SITA preferences from being stored in a struct in favour of an array
Changed the SITA preferences from being stored in a struct to an array of 4 integers. Did so as returning structs from a function requires experimental build and also trying to retrieve the struct and just returning the 4 dimension values in an array would re-order the array. Example: enter [1,2,3,4], then retrieve and it would be [1,3,2,4].